Rewrite insert nums logic completely
Now uses str.format() for formatting and allows specification of a format
string in the prompt. Furthermore, the step is now specified by ":" and may be
omitted.
Only supports numerics as of now, but can do prefixed hexadecimal 8-digit
output using "1~#010x", for example. Yeah, really powerful.

Also includes a basic syntax definition for the format syntax which I used as
base for the regular expression. Most parts are copied from the Pythons syntax
definition in order to be compatible with it.