Add stop mode (for numbers only currently)
This allows to generate only to a specific number. If less numbers are
generated than there are selections, the remaining regions are replaced
with an empty string (""). If more numbers are generated, all remaining
numbers are joined as multiline strings in the last region.

A great example is `@i>100` which generates numbers from 1 to 100.

In the process:
- Add "n" variable to eval env which represents the number of selections
- Enforce max line width of 79 characters
- Add a time limit for generating numbers in case the user entered a
  stop condition that is never true, leading to an infinite loop
- Add a "quiet" mode for preview that does not fail when an invalid
  expression was given (also for the general "expr")