Add expression mode
In this mode, indicated by a leading "|" in the format string, the
current selection will be parsed and set to `_` in the expr env.
Supports format string and stop expr, and is also able to cast the
selection to a type beforehand. Reverse mode currently does not work as
intended.

Syntax:
exprmode      ::=  [cast] "|" [~format::][expr][@stopexpr] ["!"]
cast          ::=  "i" | "f" | "s" | "b"